requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
Django==3.2.6
djangorestframework==3.12.4
fakeredis==1.6.0
flake8==3.9.2
freezegun==1.1.0
iniconfig==1.1.1
isort==5.9.3
mccabe==0.6.1
mirakuru==2.4.1
packaging==21.0
pluggy==0.13.1
port-for==0.6.1
psutil==5.8.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.1
toml==0.10.2
